chore: rename zora worker to nft
