+++ packages/data/constants.ts
@@ -107,8 +107,8 @@ export const PREFERENCES_WORKER_URL = IS_PRODUCTION
 export const CHANNELS_WORKER_URL = IS_PRODUCTION
   ? 'https://channels.hey.xyz'
   : 'http://localhost:8093';
-export const ZORA_WORKER_URL = IS_PRODUCTION
-  ? 'https://zora.hey.xyz'
+export const NFT_WORKER_URL = IS_PRODUCTION
+  ? 'https://nft.hey.xyz'
   : 'http://localhost:8094';
 
 // Tokens / Keys
