+++ packages/workers/nft/.eslintrc.js
