+++ packages/workers/nft/README.md
@@ -1,0 +1,1 @@
+# Nft worker
