+++ packages/workers/nft/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@workers/zora",
+  "name": "@workers/nft",
   "version": "0.0.0",
   "private": true,
   "license": "AGPL-3.0",
