+++ packages/workers/nft/src/constants.ts
