+++ packages/workers/nft/tsconfig.json
