+++ packages/workers/nft/wrangler.toml
@@ -1,10 +1,10 @@
-name = "zora"
+name = "nft"
 main = "src/index.ts"
 compatibility_date = "2023-01-25"
 keep_vars = true
 
 routes = [
-	{ pattern = "zora.hey.xyz", custom_domain = true }
+	{ pattern = "nft.hey.xyz", custom_domain = true }
 ]
 
 [env.production.vars]
