--- packages/workers/zora/README.md
@@ -1,1 +1,0 @@
-# ZORA worker
